% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_psi.R
\name{count_psi}
\alias{count_psi}
\title{Count PSI}
\usage{
count_psi(
  pop1_feature_vec,
  pop2_feature_vec,
  start_with_buckets = FALSE,
  ...,
  info_for_all_buckets = FALSE
)
}
\arguments{
\item{pop1_feature_vec}{Feature vector for first population.}

\item{pop2_feature_vec}{Feature vector for second population.}

\item{start_with_buckets}{Boolean value which determines whether or not
feature_vec is already bucketed. Default is FALSE, which means that
feature_vec is not bucketed.}

\item{...}{Argument no_buckets to pass to make_buckets function if
start_with_buckets is set to FALSE.}

\item{info_for_all_buckets}{Boolean value which determines
whether or not PSI will be computed for whole feature or for each bucket.
Default value is set to FALSE, which means that single value
of feature's PSI will be returned.}
}
\value{
Function returns a data frame (tibble) which contains either
a single value of feature's PSI, or (if info_for_all_buckets argument
is set to TRUE) PSI values for each bucket with
information of maximum border in buckets, etc.
}
\description{
Function counts Population Stability Index
for a given feature.
}
\examples{

a <- c(1, 1, 1, 2, 2, 3, 4, 6, 7, 8, 9, 9, 9, NA)
b <- c(NA, NA, 12, 3, 4, 5, 6, 6, 2, 2, 1, 321, 512, NA)

count_psi(a,b, no_buckets = 3)
count_psi(a,b, no_buckets = 3, info_for_all_buckets = TRUE)

# Counting PSI for multipe features
# library(tidyverse)
test_tbl \%>\% map2(test_tbl_2, count_psi)
test_tbl \%>\% map2(test_tbl_2, count_psi, no_buckets = 3, info_for_all_buckets = TRUE)

}
